+++ pom.xml
@@ -22,8 +22,10 @@
 	<properties>
 		<java.version>11</java.version>
 		<swagger.version>2.9.2</swagger.version>
-		<junit-jupiter.version>5.4.0</junit-jupiter.version>
+		<junit-jupiter.version>5.4.2</junit-jupiter.version>
+		<junit-platform.version>1.4.2</junit-platform.version>
 		<oss-java-driver.version>4.0.1</oss-java-driver.version>
+		<testcontainers.version>1.11.3</testcontainers.version>
 
 		<version.maven.plugin.compiler>3.8.0</version.maven.plugin.compiler>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -65,7 +67,6 @@
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-lang3</artifactId>
 		</dependency>
-
 		<!-- Provides JSON serialization/deserialization for date/time types -->
 		<dependency>
 			<groupId>com.fasterxml.jackson.datatype</groupId>
@@ -84,6 +85,7 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
+		<!-- Junit 5 -->
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
 			<artifactId>junit-jupiter-api</artifactId>
@@ -114,7 +116,33 @@
 			<artifactId>junit-platform-runner</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.junit.platform</groupId>
+			<artifactId>junit-platform-console-standalone</artifactId>
+			<version>${junit-platform.version}</version>
+			<scope>test</scope>
+		</dependency>
+		<!-- Test against Docker Containers -->
+		<dependency>
+			<groupId>org.testcontainers</groupId>
+			<artifactId>cassandra</artifactId>
+			<scope>test</scope>
+			<version>${testcontainers.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.testcontainers</groupId>
+			<artifactId>testcontainers</artifactId>
+			<scope>test</scope>
+			<version>${testcontainers.version}</version>
+		</dependency>
 		
+		<!-- Add driver keys to spring-boot config file -->
+		<dependency>
+           <groupId>org.springframework.boot</groupId>
+           <artifactId>spring-boot-configuration-processor</artifactId>
+           <optional>true</optional>
+        </dependency>
+        
 	</dependencies>
 
 	<build>
@@ -124,10 +152,6 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-			</plugin>
 		</plugins>
 
 		<pluginManagement>
@@ -135,7 +159,6 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
-					<version>${version.maven.plugin.compiler}</version>
 					<configuration>
 						<release>${java.version}</release>
 					</configuration>
