+++ src/main/java/com/cassandraguide/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.cassandraguide.reservation.api;
+package com.cassandraguide.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 @RequestMapping("/")
-public class HomeResource {
+public class HomeController {
     
     @RequestMapping(value = "/", method = RequestMethod.GET, produces = "text/html")
     public String redirectToDoc() {
