+++ src/test/resources/config-test.properties
@@ -5,3 +5,4 @@ cassandra.contactpoint.hostname=127.0.0.1
 cassandra.contactpoint.port=9042
 cassandra.localDataCenterName=datacenter1
 cassandra.keyspaceName=reservation
+
